cloud_functions/generate_dataform_config/requirements.txt
google-cloud-bigquery>=3.0.0
google-cloud-storage>=2.0.0
pytz
tenacity # (opcional, para reintentos)